msgr:  fix large message data content length causing overflow

+++ src/msg/simple/Pipe.h
@@ -143,9 +143,9 @@ class DispatchQueue;
     }
 
     char *recv_buf;
-    int recv_max_prefetch;
-    int recv_ofs;
-    int recv_len;
+    size_t recv_max_prefetch;
+    size_t recv_ofs;
+    size_t recv_len;
 
     enum {
       STATE_ACCEPTING,
@@ -244,7 +244,7 @@ class DispatchQueue;
      * @param more Should be set true if this is one part of a larger message
      * @return 0, or -1 on failure (unrecoverable -- close the socket).
      */
-    int do_sendmsg(struct msghdr *msg, int len, bool more=false);
+    int do_sendmsg(struct msghdr *msg, unsigned len, bool more=false);
     int write_ack(uint64_t s);
     int write_keepalive();
     int write_keepalive2(char tag, const utime_t &t);
@@ -339,8 +339,8 @@ class DispatchQueue;
       recv_len = 0;
       recv_ofs = 0;
     }
-    int do_recv(char *buf, size_t len, int flags);
-    int buffered_recv(char *buf, size_t len, int flags);
+    ssize_t do_recv(char *buf, size_t len, int flags);
+    ssize_t buffered_recv(char *buf, size_t len, int flags);
     bool has_pending_data() { return recv_len > recv_ofs; }
 
     /**
@@ -350,7 +350,7 @@ class DispatchQueue;
      * @param len exact number of bytes to read
      * @return 0 for success, or -1 on error
      */
-    int tcp_read(char *buf, int len);
+    int tcp_read(char *buf, unsigned len);
 
     /**
      * wait for bytes to become available on the socket
@@ -369,7 +369,7 @@ class DispatchQueue;
      * @param len maximum number of bytes to read
      * @return bytes read, or -1 on error or when there is no data
      */
-    int tcp_read_nonblocking(char *buf, int len);
+    ssize_t tcp_read_nonblocking(char *buf, unsigned len);
 
     /**
      * blocking write of bytes to socket
@@ -378,7 +378,7 @@ class DispatchQueue;
      * @param len number of bytes to write
      * @return 0 for success, or -1 on error
      */
-    int tcp_write(const char *buf, int len);
+    int tcp_write(const char *buf, unsigned len);
 
   };
